+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+
+## 4.4.1
+* Fix Tab issue in editing mode of `Datagrid`
+* Add align to input components of `Datagrid` if prop `align` is set
+
+## 4.4.0
 * Add `FloatingSelect` and `FloatingSelectCreatable` components
 * Fix datagrid tooltip position